// src/model/proto/credential/account.credential.grpc.ts
export interface IAccountCredentialParent {
  /** Document ID (_id in MongoDB) */
  id?:
    | string
    | undefined;
  /** Preference ObjectId (as string) */
  credential?:
    | string
    | undefined;
  /** Preference ObjectId (as string) */
  info?: string | undefined;
}

/** Defining the AccountCredential message based on the IAccountCredential model */
export interface IAccountCredential {
  /** Document ID (_id in MongoDB) */
  id?:
    | string
    | undefined;
  /** Preference ObjectId (as string) */
  preference?:
    | string
    | undefined;
  /** Parent ObjectId (as string) */
  parent?:
    | IAccountCredentialParent
    | undefined;
  /** First name as string */
  email: string;
  /** Last name as string */
  username: string;
  password?: string | undefined;
}

/** The response message containing an AccountCredential */
export interface AccountCredentialReadRequest {
}

export interface AccountCredentialAuthRequest {
  username: string;
  password: string;
}

export interface AccountCredentialAuthResponse {
  status: boolean;
  code: number;
  msg: string;
  data: IAccountCredential | undefined;
}

export interface AccountCredentialCreateResponse {
  status: boolean;
  code: number;
  msg: string;
  data: IAccountCredential | undefined;
}

export interface AccountCredentialReadResponse {
  status: boolean;
  code: number;
  msg: string;
  data: IAccountCredential[];
}
// ...
export const IAccountCredential: MessageFns<IAccountCredential> = {
  encode(message: IAccountCredential, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== undefined) {
      writer.uint32(10).string(message.id);
    }
    if (message.preference !== undefined) {
      writer.uint32(18).string(message.preference);
    }
    if (message.parent !== undefined) {
      IAccountCredentialParent.encode(message.parent, writer.uint32(26).fork()).join();
    }
    if (message.email !== "") {
      writer.uint32(34).string(message.email);
    }
    if (message.username !== "") {
      writer.uint32(42).string(message.username);
    }
    if (message.password !== undefined) {
      writer.uint32(50).string(message.password);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): IAccountCredential {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseIAccountCredential();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.preference = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.parent = IAccountCredentialParent.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.email = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.username = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.password = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): IAccountCredential {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : undefined,
      preference: isSet(object.preference) ? globalThis.String(object.preference) : undefined,
      parent: isSet(object.parent) ? IAccountCredentialParent.fromJSON(object.parent) : undefined,
      email: isSet(object.email) ? globalThis.String(object.email) : "",
      username: isSet(object.username) ? globalThis.String(object.username) : "",
      password: isSet(object.password) ? globalThis.String(object.password) : undefined,
    };
  },

  toJSON(message: IAccountCredential): unknown {
    const obj: any = {};
    if (message.id !== undefined) {
      obj.id = message.id;
    }
    if (message.preference !== undefined) {
      obj.preference = message.preference;
    }
    if (message.parent !== undefined) {
      obj.parent = IAccountCredentialParent.toJSON(message.parent);
    }
    if (message.email !== "") {
      obj.email = message.email;
    }
    if (message.username !== "") {
      obj.username = message.username;
    }
    if (message.password !== undefined) {
      obj.password = message.password;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<IAccountCredential>, I>>(base?: I): IAccountCredential {
    return IAccountCredential.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<IAccountCredential>, I>>(object: I): IAccountCredential {
    const message = createBaseIAccountCredential();
    message.id = object.id ?? undefined;
    message.preference = object.preference ?? undefined;
    message.parent = (object.parent !== undefined && object.parent !== null)
      ? IAccountCredentialParent.fromPartial(object.parent)
      : undefined;
    message.email = object.email ?? "";
    message.username = object.username ?? "";
    message.password = object.password ?? undefined;
    return message;
  },
};
// ...
export const AccountCredentialCreateResponse: MessageFns<AccountCredentialCreateResponse> = {
  encode(message: AccountCredentialCreateResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.status !== false) {
      writer.uint32(8).bool(message.status);
    }
    if (message.code !== 0) {
      writer.uint32(16).int64(message.code);
    }
    if (message.msg !== "") {
      writer.uint32(26).string(message.msg);
    }
    if (message.data !== undefined) {
      IAccountCredential.encode(message.data, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AccountCredentialCreateResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAccountCredentialCreateResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.status = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.code = longToNumber(reader.int64());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.msg = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.data = IAccountCredential.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AccountCredentialCreateResponse {
    return {
      status: isSet(object.status) ? globalThis.Boolean(object.status) : false,
      code: isSet(object.code) ? globalThis.Number(object.code) : 0,
      msg: isSet(object.msg) ? globalThis.String(object.msg) : "",
      data: isSet(object.data) ? IAccountCredential.fromJSON(object.data) : undefined,
    };
  },

  toJSON(message: AccountCredentialCreateResponse): unknown {
    const obj: any = {};
    if (message.status !== false) {
      obj.status = message.status;
    }
    if (message.code !== 0) {
      obj.code = Math.round(message.code);
    }
    if (message.msg !== "") {
      obj.msg = message.msg;
    }
    if (message.data !== undefined) {
      obj.data = IAccountCredential.toJSON(message.data);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AccountCredentialCreateResponse>, I>>(base?: I): AccountCredentialCreateResponse {
    return AccountCredentialCreateResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AccountCredentialCreateResponse>, I>>(
    object: I,
  ): AccountCredentialCreateResponse {
    const message = createBaseAccountCredentialCreateResponse();
    message.status = object.status ?? false;
    message.code = object.code ?? 0;
    message.msg = object.msg ?? "";
    message.data = (object.data !== undefined && object.data !== null)
      ? IAccountCredential.fromPartial(object.data)
      : undefined;
    return message;
  },
};

function createBaseAccountCredentialReadResponse(): AccountCredentialReadResponse {
  return { status: false, code: 0, msg: "", data: [] };
}

export const AccountCredentialReadResponse: MessageFns<AccountCredentialReadResponse> = {
  encode(message: AccountCredentialReadResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.status !== false) {
      writer.uint32(8).bool(message.status);
    }
    if (message.code !== 0) {
      writer.uint32(16).int64(message.code);
    }
    if (message.msg !== "") {
      writer.uint32(26).string(message.msg);
    }
    for (const v of message.data) {
      IAccountCredential.encode(v!, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AccountCredentialReadResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAccountCredentialReadResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.status = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.code = longToNumber(reader.int64());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.msg = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.data.push(IAccountCredential.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AccountCredentialReadResponse {
    return {
      status: isSet(object.status) ? globalThis.Boolean(object.status) : false,
      code: isSet(object.code) ? globalThis.Number(object.code) : 0,
      msg: isSet(object.msg) ? globalThis.String(object.msg) : "",
      data: globalThis.Array.isArray(object?.data) ? object.data.map((e: any) => IAccountCredential.fromJSON(e)) : [],
    };
  },

  toJSON(message: AccountCredentialReadResponse): unknown {
    const obj: any = {};
    if (message.status !== false) {
      obj.status = message.status;
    }
    if (message.code !== 0) {
      obj.code = Math.round(message.code);
    }
    if (message.msg !== "") {
      obj.msg = message.msg;
    }
    if (message.data?.length) {
      obj.data = message.data.map((e) => IAccountCredential.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AccountCredentialReadResponse>, I>>(base?: I): AccountCredentialReadResponse {
    return AccountCredentialReadResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AccountCredentialReadResponse>, I>>(
    object: I,
  ): AccountCredentialReadResponse {
    const message = createBaseAccountCredentialReadResponse();
    message.status = object.status ?? false;
    message.code = object.code ?? 0;
    message.msg = object.msg ?? "";
    message.data = object.data?.map((e) => IAccountCredential.fromPartial(e)) || [];
    return message;
  },
};